Reworked jobs. Made two types of Jobs: Creation jobs and upgrade Jobs. Creation Jobs accept a Unit *type* (not an instance of Unit). Upon completion of a creation job, a unit is instantiated. Upgrade jobs accept an existing unit instance. Upon completion of the job, the unit is automatically upgraded to the next level. Made unit attributes into *class* attributes to make it so that jobs can easily instantiate units without passing in arguments. Changed classes in unit_classes/ to reflect these changes

+++ engine/location.py
@@ -31,7 +31,7 @@ class Path(GameObject):
 
         self._distance = GameNode.distance(self._game_node1, self._game_node2)
 
-        self._armies_and_coords: dict[Army, int] = {}  # armies and their corresponding position on the path
+        self._armies_and_coords: dict[Army, float] = {}  # armies and their corresponding position on the path
 
     @property
     def distance(self) -> float | int:
